requirements.txt
# Core dependencies
PyYAML>=6.0

# Development dependencies
pytest>=7.0
pytest-cov>=4.0
black>=23.0
flake8>=6.0
mypy>=1.0
